+++ src/r_effects.cc
@@ -157,10 +157,10 @@ void RGL_ColourmapEffect(player_t *player)
 		y1 = viewwindow_y + viewwindow_h;
 		y2 = viewwindow_y;
 
-		buffer.add({x1, y1, 0});
-		buffer.add({x2, y1, 0});
-		buffer.add({x2, y2, 0});
-		buffer.add({x1, y2, 0});
+		buffer.add({(float)x1, (float)y1, 0});
+		buffer.add({ (float)x2, (float)y1, 0});
+		buffer.add({ (float)x2, (float)y2, 0});
+		buffer.add({ (float)x1, (float)y2, 0});
 
 		buffer.draw(GL_QUADS);
 	  
@@ -244,9 +244,9 @@ void RGL_PaletteEffect(player_t *player)
 	glEnable(GL_BLEND);
 
 	RQImmBuffer<RQVertex3f> buffer(RQVertex3f::format);
-	buffer.add({0, SCREENHEIGHT, 0});
-	buffer.add({SCREENWIDTH, SCREENHEIGHT, 0});
-	buffer.add({SCREENWIDTH, 0, 0});
+	buffer.add({ (float)0, (float)SCREENHEIGHT, 0});
+	buffer.add({ (float)SCREENWIDTH, (float)SCREENHEIGHT, 0});
+	buffer.add({ (float)SCREENWIDTH, (float)0, 0});
 	buffer.add({0, 0, 0});
 	buffer.draw(GL_QUADS);
